+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - support AUTO directive, highlight as unsupported
+- support MSOURCE directive
 
 ### Fixed
 - Show used/free registers: handle register-list/range, PC indirect with index